Move create user to new file
