+++ src/components/SignUpForm/SignUpForm.js
@@ -4,6 +4,7 @@ import request from 'superagent';
 import { useHistory, Redirect } from "react-router-dom";
 import { useDispatch, useSelector } from 'react-redux';
 import { setNewUser, loginCurrentUser } from '../../actions';
+import { createUser } from '../../utils/usersAPICalls';
 import Loader from '../Loader/Loader';
 import './SignUpForm.scss';
 
@@ -66,22 +67,7 @@ const SignUpForm = () => {
       };
     }
 
-    const options = {
-      method: 'POST',
-      body: JSON.stringify(mutation),
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      redirect: 'follow'
-    };
-
-    return fetch('https://ican2-be-rails.herokuapp.com/api/v1/graphql', options)
-    .then(response => {
-      if (!response.ok) {
-          throw Error('error retrieving user data');
-        }
-      return response.json();
-    });
+    return createUser(mutation);
   };
 
   const login = (e) => {
